refactor: remove components/pages/ directory
