+++ frontend/src/app/login/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Login from "@/components/pages/auth/Login";
+import Login from "@/components/auth/Login";
 
 const LoginPage = () => {
   return <Login />;
