+++ frontend/src/app/page.tsx
@@ -1,4 +1,4 @@
-import Home from "@/components/pages/Home";
+import Home from "@/components/Home";
 
 const HomePage = () => {
   return <Home />;
