+++ frontend/src/components/auth/Login.tsx
@@ -70,7 +70,7 @@ const Login = (props: Props) => {
         <div className="mt-4 text-center">
           <p className="text-sm">
             Don&apos;t have an account?{" "}
-            <a href="/signup" className="text-blue-600 hover:underline">
+            <a href="/frontend/src/components/auth/Signup" className="text-blue-600 hover:underline">
               Sign up
             </a>
           </p>
