+++ frontend/src/components/event/eventDetail.tsx
